requirements.txt
numpy
matplotlib
Pillow
streamlit
streamlit_drawable_canvas
scipy
jupyter
notebook
nbconvert
tk
